+++ BackendWebAPI/Controllers/LoginController.cs
@@ -1,15 +1,20 @@
 ﻿using BackendWebAPI.Models;
 using BackendWebAPI.Services;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
 namespace BackendWebAPI.Controllers
 {
     public class LoginController : Controller
     {
-        private readonly PixelWarsService _pixelwarsService;
-        public LoginController(PixelWarsService pixelwarsService) =>
-            _pixelwarsService = pixelwarsService;
+       /* private readonly SignInManager<Person> _signInManager;
+        public LoginController(SignInManager<Person> signInManager)
+        {
+            _signInManager = signInManager;
+        }*/
 
         [Route("/login")]
         [HttpGet]
@@ -18,66 +23,12 @@ namespace BackendWebAPI.Controllers
             return View();
         }
 
-        [Route("/")]
-        [Authorize]
+       /* [Route("/logout")]
         [HttpGet]
-        public IActionResult Main()
+        public async Task<ActionResult> Logout()
         {
-            return View();
-        }
-
-        [Route("/CreateGrid")]
-        [Authorize]
-        [HttpGet]
-        public IActionResult CreateGrid()
-        {
-            return View();
-        }
-
-        [Route("/InsertGrid")]
-        [Authorize]
-        [HttpPost]
-        public async Task<ActionResult<PixelWarsCollection>> InsertGrid(string name, string size)
-        {
-            var insertValues = new PixelWarsCollection();
-            insertValues.Name = name;
-            insertValues.Size = size;
-
-            await _pixelwarsService.InsertGridBynameAndSize(insertValues);
-
-            var filter = await _pixelwarsService.GetByNameAsync(name);
-
-            if (filter is null)
-            {
-                return NotFound();
-            }
-
-            return View();
-        }
-
-        [Route("/DeleteGrid")]
-        [Authorize]
-        [HttpGet]
-        public IActionResult DeleteGrid()
-        {
-            return View();
-        }
-
-        [Route("/DeleteGridbyName")]
-        [Authorize]
-        [HttpPost]
-        public async Task<ActionResult<PixelWarsCollection>> DeleteGridbyName(string name)
-        {
-            await _pixelwarsService.DeleteGridByName(name);
-
-            var filter = await _pixelwarsService.GetByNameAsync(name);
-
-            if (filter is null)
-            {
-                return View();
-            }
-
-            return NotFound();
-        }
+            await _signInManager.SignOutAsync();
+            return Redirect("/login");
+        }*/
     }
 }
