+++ BackendWebAPI/Views/Login/Index.cshtml
@@ -1,9 +1,5 @@
 ﻿
 <div class="login">
-    <div id="userInfo" style="display:none;">
-        <p>Добро пожаловать <span id="userName"></span>!</p>
-        <input type="button" value="Выйти" id="logOut" />
-    </div>
     <div id="loginForm">
         <h1>Авторизация</h1>
         <form method="post" enctype="multipart/form-data">
