+++ BackendWebAPI/Views/Shared/_Layout.cshtml
@@ -10,36 +10,26 @@
     <link rel="stylesheet" href="~/css/site.css" asp-append-version="true" />
     <link href="https://fonts.cdnfonts.com/css/common-pixel" rel="stylesheet">
 </head>
-<body>
+
+<header>
     @if (HttpContextAccessor.HttpContext.User.Identity.IsAuthenticated)
     {
-        <div class='left_side_header'>
-            <span class="logo">PixelWars</span>
-            <p>
-                <ul id="navbar">
-                    <li>
-                        <a href="/">Главная</a>
-                    </li>
-                    <li>
-                        <a href="/CreateGrid">Создать Грид</a>
-                    </li>
-                    <li>
-                        <a href="/DeleteGrid">Удалить Грид</a>
-                    </li>
-                </ul>
-            </p>
-        </div>
+        <a href="/"> <span class="logo">PixelWars</span></a>
 
-        <div class='right_top_button'>
-            <div class="logout">
-                <a href="/logout"><p>Выход</p></a>
-            </div>
+        <navbar style="flex:auto;">            
+                <a href="/CreateGrid">Создать Грид</a>  
+                <a href="/DeleteGrid">Удалить Грид</a>            
+        </navbar>
+
+        <div class="logout">
+            <a href="/logout"><p>Выход</p></a>
         </div>
     }
+</header>
 
+<body>
     <main>
         @RenderBody()
     </main>
-
 </body>
 </html>
