+++ BackendWebAPI/wwwroot/css/site.css
@@ -1,5 +1,21 @@
 ﻿@import url('https://fonts.cdnfonts.com/css/common-pixel');
 
+a:link {
+    text-decoration: none;
+}
+
+a:visited {
+    text-decoration: none;
+}
+
+a:hover {
+    text-decoration: none;
+}
+
+a:active {
+    text-decoration: none;
+}
+
 body {
     font: 13px/20px "Lucida Grande", Tahoma, Verdana, sans-serif;
     color: #000000;
@@ -21,7 +37,7 @@ body {
     width: 45px;
     height: 20px;
     display: flex;
-    justify-content: center;
+    justify-content: flex-end;
     align-items: center;
 }
